+++ src/lib/modlist/ModListItem.svelte
@@ -2,7 +2,6 @@
 	import type { Mod } from '../models';
 	import Icon from '@iconify/svelte';
 	import { isOutdated } from '$lib/util';
-	import { invokeCommand } from '$lib/invoke';
 
 	const FALLBACK_ICON =
 		'https://sm.ign.com/t/ign_es/cover/l/lethal-com/lethal-company_817h.300.jpg';
@@ -11,14 +10,6 @@
 	export let isSelected: boolean;
 	export let showInstalledIcon: boolean;
 	export let draggable = false;
-
-	let isInstalled = false;
-
-	$: refreshInstalled(mod);
-
-	async function refreshInstalled(mod: Mod) {
-		isInstalled = await invokeCommand('is_mod_installed', { uuid: mod.uuid });
-	}
 </script>
 
 <button
@@ -47,7 +38,7 @@
 		>
 			{mod.version ?? ''}
 		</span>
-		{#if isInstalled && showInstalledIcon}
+		{#if false && showInstalledIcon}
 			<Icon class="text-green-500 inline mb-1" icon="mdi:check" />
 		{/if}
 		{#if mod.isDeprecated}
@@ -65,6 +56,6 @@
 
 	<!-- svelte-ignore a11y-click-events-have-key-events -->
 	<div class="contents" on:click={(evt) => evt.stopPropagation()} role="none">
-		<slot {isInstalled} />
+		<slot isInstalled={false} />
 	</div>
 </button>